resolve conflicts, adopt tests to changes
